fixing according to the code review
